refactor: extract pass XP logic to its own method
